Add tree to Store and move Store constructor
