+++ lib/types/store.ex
@@ -2,6 +2,17 @@ defmodule Types.Store do
   @moduledoc """
     The Store struct is used to track information required for the fork choice algorithm.
   """
+
+  alias LambdaEthereumConsensus.ForkChoice.Tree
+  alias LambdaEthereumConsensus.StateTransition.Misc
+  alias LambdaEthereumConsensus.Store.Blocks
+  alias LambdaEthereumConsensus.Store.StateStore
+  alias LambdaEthereumConsensus.StateTransition.Accessors
+  alias Types.BeaconBlock
+  alias Types.BeaconState
+  alias Types.Checkpoint
+  alias Types.SignedBeaconBlock
+
   defstruct [
     :time,
     :genesis_time,
@@ -15,31 +26,72 @@ defmodule Types.Store do
     :block_states,
     :checkpoint_states,
     :latest_messages,
-    :unrealized_justifications
+    :unrealized_justifications,
+    # Stores block data on the current fork tree (~last two epochs)
+    :tree_cache
   ]
 
   @type t :: %__MODULE__{
           time: Types.uint64(),
           genesis_time: Types.uint64(),
-          justified_checkpoint: Types.Checkpoint.t() | nil,
-          finalized_checkpoint: Types.Checkpoint.t(),
-          unrealized_justified_checkpoint: Types.Checkpoint.t() | nil,
-          unrealized_finalized_checkpoint: Types.Checkpoint.t() | nil,
+          justified_checkpoint: Checkpoint.t() | nil,
+          finalized_checkpoint: Checkpoint.t(),
+          unrealized_justified_checkpoint: Checkpoint.t() | nil,
+          unrealized_finalized_checkpoint: Checkpoint.t() | nil,
           proposer_boost_root: Types.root() | nil,
           equivocating_indices: MapSet.t(Types.validator_index()),
-          blocks: %{Types.root() => Types.BeaconBlock.t()},
-          block_states: %{Types.root() => Types.BeaconState.t()},
-          checkpoint_states: %{Types.Checkpoint.t() => Types.BeaconState.t()},
-          latest_messages: %{Types.validator_index() => Types.Checkpoint.t()},
-          unrealized_justifications: %{Types.root() => Types.Checkpoint.t()}
+          blocks: %{Types.root() => BeaconBlock.t()},
+          block_states: %{Types.root() => BeaconState.t()},
+          checkpoint_states: %{Checkpoint.t() => BeaconState.t()},
+          latest_messages: %{Types.validator_index() => Checkpoint.t()},
+          unrealized_justifications: %{Types.root() => Checkpoint.t()},
+          tree_cache: Tree.t()
         }
 
-  alias LambdaEthereumConsensus.StateTransition.Misc
-  alias LambdaEthereumConsensus.Store.Blocks
-  alias LambdaEthereumConsensus.Store.BlockStore
-  alias LambdaEthereumConsensus.Store.StateStore
-  alias Types.BeaconState
-  alias Types.SignedBeaconBlock
+  @spec get_forkchoice_store(BeaconState.t(), SignedBeaconBlock.t(), boolean()) ::
+          {:ok, Store.t()} | {:error, String.t()}
+  def get_forkchoice_store(
+        %BeaconState{} = anchor_state,
+        %SignedBeaconBlock{message: anchor_block} = signed_block,
+        use_db
+      ) do
+    anchor_state_root = Ssz.hash_tree_root!(anchor_state)
+    anchor_block_root = Ssz.hash_tree_root!(anchor_block)
+
+    if anchor_block.state_root == anchor_state_root do
+      anchor_epoch = Accessors.get_current_epoch(anchor_state)
+
+      anchor_checkpoint = %Checkpoint{
+        epoch: anchor_epoch,
+        root: anchor_block_root
+      }
+
+      time = anchor_state.genesis_time + ChainSpec.get("SECONDS_PER_SLOT") * anchor_state.slot
+
+      %__MODULE__{
+        time: time,
+        genesis_time: anchor_state.genesis_time,
+        justified_checkpoint: anchor_checkpoint,
+        finalized_checkpoint: anchor_checkpoint,
+        unrealized_justified_checkpoint: anchor_checkpoint,
+        unrealized_finalized_checkpoint: anchor_checkpoint,
+        proposer_boost_root: <<0::256>>,
+        equivocating_indices: MapSet.new(),
+        blocks: %{},
+        block_states: %{},
+        checkpoint_states: %{anchor_checkpoint => anchor_state},
+        latest_messages: %{},
+        unrealized_justifications: %{anchor_block_root => anchor_checkpoint},
+        tree_cache: Tree.new(anchor_block_root)
+      }
+      |> then(&if use_db, do: &1 |> Map.delete(:blocks) |> Map.delete(:block_states), else: &1)
+      |> store_block(anchor_block_root, signed_block)
+      |> store_state(anchor_block_root, anchor_state)
+      |> then(&{:ok, &1})
+    else
+      {:error, "Anchor block state root does not match anchor state root"}
+    end
+  end
 
   def get_current_slot(%__MODULE__{time: time, genesis_time: genesis_time}) do
     # NOTE: this assumes GENESIS_SLOT == 0
@@ -98,12 +150,12 @@ defmodule Types.Store do
     store
   end
 
-  @spec get_block(t(), Types.root()) :: Types.BeaconBlock.t() | nil
+  @spec get_block(t(), Types.root()) :: BeaconBlock.t() | nil
   def get_block(%__MODULE__{blocks: blocks}, block_root) do
     Map.get(blocks, block_root)
   end
 
-  @spec get_block(t(), Types.root()) :: Types.BeaconBlock.t() | nil
+  @spec get_block(t(), Types.root()) :: BeaconBlock.t() | nil
   def get_block(%__MODULE__{}, block_root) do
     case Blocks.get_block(block_root) do
       nil -> nil
@@ -111,7 +163,7 @@ defmodule Types.Store do
     end
   end
 
-  @spec get_block!(t(), Types.root()) :: Types.BeaconBlock.t()
+  @spec get_block!(t(), Types.root()) :: BeaconBlock.t()
   def get_block!(store, block_root) do
     case get_block(store, block_root) do
       nil -> raise "Block not found: 0x#{Base.encode16(block_root)}"
@@ -119,9 +171,9 @@ defmodule Types.Store do
     end
   end
 
-  @spec get_blocks(t()) :: Enumerable.t(Types.BeaconBlock.t())
+  @spec get_blocks(t()) :: Enumerable.t(BeaconBlock.t())
   def get_blocks(%__MODULE__{blocks: blocks}), do: blocks
-  def get_blocks(%__MODULE__{}), do: BlockStore.stream_blocks()
+  def get_blocks(%__MODULE__{tree_cache: tree}), do: Tree.get_all_blocks(tree)
 
   @spec store_block(t(), Types.root(), SignedBeaconBlock.t()) :: t()
   def store_block(%__MODULE__{blocks: blocks} = store, block_root, %{message: block}) do
